created lexerToken.js removed all global variables and restructured the code
