+++ lib/tokens.js
@@ -92,5 +92,4 @@ for(let i = 1; i < inbuiltFunctionList.length; i++) temporaryRegex += "|" + inbu
 // console.log(temporaryRegex);
 tokens.inbuilt_function.match = temporaryRegex;
 
-
 exports.tokens = tokens;
